Reduce smoke test suite size
The smoke tests are taking too long, 5m on linux and 10m+ on windows on
our CI servers. The goal for the smoke tests is <3m so that it's easy to
run and vet that you didn't completely break porter.

I've moved some of the smoke tests into the integration test suite so
that we are only focusing on the most important checks in the smoke test
suite.
