+++ tests/integration/archive_test.go
@@ -1,5 +1,4 @@
 //go:build integration
-// +build integration
 
 package integration
 
@@ -14,7 +13,12 @@ import (
 
 	"get.porter.sh/porter/pkg"
 	"get.porter.sh/porter/pkg/porter"
+	"get.porter.sh/porter/pkg/yaml"
 	"get.porter.sh/porter/tests"
+	"get.porter.sh/porter/tests/testdata"
+	"get.porter.sh/porter/tests/tester"
+	"github.com/carolynvs/magex/mgx"
+	"github.com/carolynvs/magex/shx"
 	"github.com/cnabio/cnab-go/bundle/loader"
 	"github.com/cnabio/cnab-go/packager"
 	"github.com/cnabio/cnab-to-oci/relocation"
@@ -22,7 +26,48 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestArchive(t *testing.T) {
+// Validate that when we archive a bundle, everything is included inside
+// We should be able to archive, move to a disconnected environment
+// and publish again without accidentally referencing artifacts from the original environment.
+func TestArchive_AirgappedEnvironment(t *testing.T) {
+	test, err := tester.NewTest(t)
+	defer test.Close()
+	require.NoError(t, err, "test setup failed")
+
+	// Start a temp registry
+	tempRegistryId, err := shx.OutputE("docker", "run", "-d", "-P", "registry:2")
+	require.NoError(t, err, "Could not start a temporary registry")
+	stopTempRegistry := func() error {
+		return shx.RunE("docker", "rm", "-vf", tempRegistryId)
+	}
+	defer stopTempRegistry()
+
+	// Get the port that it is running on
+	tempRegistryPort, err := shx.OutputE("docker", "inspect", tempRegistryId, "--format", `{{ (index (index .NetworkSettings.Ports "5000/tcp") 0).HostPort }}`)
+	require.NoError(t, err, "Could not get the published port of the temporary registry")
+
+	// Publish referenced bundle to one location
+	localRegRef := fmt.Sprintf("localhost:%s/whalesayd:latest", tempRegistryPort)
+	require.NoError(t, shx.RunE("docker", "pull", "carolynvs/whalesayd@sha256:8b92b7269f59e3ed824e811a1ff1ee64f0d44c0218efefada57a4bebc2d7ef6f"))
+	require.NoError(t, shx.RunE("docker", "tag", "carolynvs/whalesayd@sha256:8b92b7269f59e3ed824e811a1ff1ee64f0d44c0218efefada57a4bebc2d7ef6f", localRegRef))
+	require.NoError(t, shx.RunE("docker", "push", localRegRef))
+
+	// publish a test bundle that reference the image from the temp registry
+	originTestBun := filepath.Join(test.RepoRoot, fmt.Sprintf("tests/testdata/%s/porter.yaml", testdata.MyBunsWithImgReference))
+	testBun := filepath.Join(test.TestDir, "mybuns-img-reference.yaml")
+	mgx.Must(shx.Copy(originTestBun, testBun))
+	test.EditYaml(testBun, func(yq *yaml.Editor) error {
+		return yq.SetValue("images.whalesayd.repository", fmt.Sprintf("localhost:%s/whalesayd", tempRegistryPort))
+	})
+	test.RequirePorter("publish", "--file", "mybuns-img-reference.yaml", "--dir", test.TestDir)
+	stopTempRegistry()
+
+	archiveFilePath := filepath.Join(test.TestDir, "archive-test.tgz")
+	test.RequirePorter("archive", archiveFilePath, "--reference", testdata.MyBunsWithImgReferenceRef)
+}
+
+// Validate that archiving a bundle twice results in the same digest
+func TestArchive_StableDigest(t *testing.T) {
 	t.Parallel()
 
 	p := porter.NewTestPorter(t)
