+++ tests/smoke/hello_test.go
@@ -1,5 +1,4 @@
 //go:build smoke
-// +build smoke
 
 package smoke
